[remark-abbr] Migrate to micromark
This implementation is pretty much taken verbatim from
https://github.com/richardTowers/remark-abbr. I didn't make much of an
attempt to merge it with the existing code.

The existing test suite is retained, and I haven't moved any of the
tests over from richardTowers/remark-abbr yet.

A couple of the snapshots needed to be updated, but in my view what the
code does now is more correct than what was happening before, so I think
they're okay.

Co-Authored-By: Stalone <stalone@boxph.one>